fix: determine pointer-event for dragging based on shape kind
