+++ packages/core/src/renderer/Renderer.ts
@@ -3,6 +3,7 @@ import { Shape } from "types/shape";
 import { dragUpdate } from "./dragUtils";
 import { IStrV } from "types/value";
 import { LabelCache, State } from "types/state";
+import { isLinelike, isRectlike } from "renderer/ShapeDef";
 
 export interface ShapeProps {
   shape: Shape;
@@ -67,7 +68,15 @@ export const DraggableShape = (
     canvasSize: canvasSizeCustom ? canvasSizeCustom : shapeProps.canvasSize,
   });
   const g = document.createElementNS("http://www.w3.org/2000/svg", "g");
-  g.setAttribute("pointer-events", "bounding-box");
+  const { shapeType } = shapeProps.shape;
+  if (isLinelike(shapeType)) {
+    console.log(shapeProps.shape);
+    g.setAttribute("pointer-events", "visibleStroke");
+  } else if (isRectlike(shapeType)) {
+    g.setAttribute("pointer-events", "bounding-box");
+  } else {
+    g.setAttribute("pointer-events", "auto");
+  }
   g.appendChild(elem);
 
   const onMouseDown = (e: MouseEvent) => {
